+++ gecko-fit/src/components/header.js
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+
+class Header extends Component {
+    render() {
+      return (
+        <header className="App__header">
+          <h1 className="App__title">Gecko Fit</h1>
+        </header>
+      );
+    }
+  }
+
+export default Header;
